+++ src/index.html
@@ -32,7 +32,6 @@
   </section>
   <div id="modal"></div>
 
-  <script src="data.js"></script>
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
